+++ include/roq/logging.hpp
@@ -32,7 +32,7 @@ static void helper(roq::logging::Level log_level, roq::format_str<Args...> const
 #endif
   fmt::format_to(std::back_inserter(message), "L{} {}:{}] "sv, level, fmt.file_name, fmt.line);
   fmt::vformat_to(std::back_inserter(message), fmt.str, fmt::make_format_args(args...));
-  (*roq::logging::Handler::INSTANCE)(log_level, message);
+  roq::logging::Handler::get_instance()(log_level, message);
 }
 
 #ifndef NDEBUG
@@ -43,7 +43,7 @@ static void helper_debug(roq::logging::Level log_level, roq::format_str<Args...>
   message.clear();  // note! capacity is in reality preserved but it is not guaranteed by the standard
   fmt::format_to(std::back_inserter(message), "L{} {}:{}] DEBUG: "sv, level, fmt.file_name, fmt.line);
   fmt::vformat_to(std::back_inserter(message), fmt.str, fmt::make_format_args(args...));
-  (*roq::logging::Handler::INSTANCE)(log_level, message);
+  roq::logging::Handler::get_instance()(log_level, message);
 }
 #endif
 
@@ -54,7 +54,7 @@ static void helper_system_error(roq::logging::Level log_level, int error, roq::f
   message.clear();
   fmt::format_to(std::back_inserter(message), "L{} {}:{}] {} [{}] "sv, level, fmt.file_name, fmt.line, std::strerror(error), error);
   fmt::vformat_to(std::back_inserter(message), fmt.str, fmt::make_format_args(args...));
-  (*roq::logging::Handler::INSTANCE)(log_level, message);
+  roq::logging::Handler::get_instance()(log_level, message);
 }
 }  // namespace detail